wordSuggesterのファイル分割
